# More efficient parsing of optional values