Codecs generated by zio-json for case classes have inefficiency that during parsing for each optional field a new instance of `JsonDecoder.option` is created that allocates a small array of chars and a `Some` instance.

Both parsing methods from strings and ASTs were improved, but I have benchmarks for parsing from strings only.

Below are throughput results for real-world message samples using Scala 3 and JDK 21 on Intel® Core™ i7-11800H CPU @ 2.3GHz (max 4.6GHz).

#### Before
```txt
Benchmark                        (size)   Mode  Cnt        Score       Error  Units
GeoJSONReading.zioJson              N/A  thrpt    5     7776.653 ±   216.260  ops/s
GitHubActionsAPIReading.zioJson     N/A  thrpt    5   275023.115 ±  7134.200  ops/s
GoogleMapsAPIReading.zioJson        N/A  thrpt    5    17293.129 ±   620.785  ops/s
OpenRTBReading.zioJson              N/A  thrpt    5   141680.282 ±  4050.499  ops/s
TwitterAPIReading.zioJson           N/A  thrpt    5    14615.513 ±   869.314  ops/s
```

#### After
```txt
Benchmark                        (size)   Mode  Cnt        Score       Error  Units
GeoJSONReading.zioJson              N/A  thrpt    5     7915.853 ±   204.287  ops/s
GitHubActionsAPIReading.zioJson     N/A  thrpt    5   278922.406 ±   847.529  ops/s
GoogleMapsAPIReading.zioJson        N/A  thrpt    5    18011.900 ±  1391.679  ops/s
OpenRTBReading.zioJson              N/A  thrpt    5   152497.736 ±  3224.405  ops/s
TwitterAPIReading.zioJson           N/A  thrpt    5    15905.385 ±   500.720  ops/s
```